+++ lib/dassets/runner.rb
@@ -24,8 +24,12 @@ class Dassets::Runner
       abs_paths = @cmd_args.map{ |path| File.expand_path(path, @pwd) }
       Dassets::Cmds::DigestCmd.for(abs_paths).run
     when 'cache'
-      require 'dassets/runner/cache_command'
-      CacheCommand.new(@cmd_args.first).run
+      require 'dassets/cmds/cache_cmd'
+      cache_root_path = File.expand_path(@cmd_args.first, @pwd)
+      unless cache_root_path && File.directory?(cache_root_path)
+        raise CmdError, "specify an existing cache directory"
+      end
+      Dassets::Cmds::CacheCmd.new(cache_root_path).run
     when 'null'
       NullCommand.new.run
     else
